# apps/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  dto.StageRequest:
    properties:
      manga:
        type: string
      name:
        type: string
      order:
        type: integer
      translations:
        additionalProperties:
          $ref: '#/definitions/dto.StageTranslationRequest'
        type: object
    type: object
  dto.StageResponse:
    properties:
      description:
        type: string
      id:
        type: string
      manga:
        type: string
      name:
        type: string
      order:
        type: integer
      picture:
        type: string
      pictureCard:
        type: string
      pictureLqip:
        type: string
      pictureStatus:
        type: string
      pictureThumb:
        type: string
    type: object
  dto.StageTranslationRequest:
    properties:
      description:
        type: string
    type: object
  dto.StageTranslationResponse:
    properties:
      description:
        type: string
    type: object
  dto.StageTranslationsResponse:
    properties:
      translations:
        additionalProperties:
          $ref: '#/definitions/dto.StageTranslationResponse'
        type: object
    type: object
...
info:
  contact: {}
  description: Backend API for JojoOnePieceSimulator2 - Google-only auth, Stands and
    Devil Fruits catalogues.
  title: JojoOnePieceSimulator2 API
  version: "1.0"
paths:
...
  /stages:
    get:
      description: Lists every Stage, or filters them if any query param is set. Description
        resolved for the request's locale.
      parameters:
      - description: JOJO, ONE_PIECE
        in: query
        name: manga
        type: string
      - description: free-text search over name and description
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.StageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List or filter stages
      tags:
      - stages
    post:
      consumes:
      - application/json
      description: Admin only.
      parameters:
      - description: Stage to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.StageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a stage
      tags:
      - stages
  /stages/{id}:
    delete:
      description: Admin only.
      parameters:
      - description: Stage id (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a stage
      tags:
      - stages
    get:
      description: Description resolved for the request's locale.
      parameters:
      - description: Stage id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a stage by id
      tags:
      - stages
    put:
      consumes:
      - application/json
      description: Admin only. Keeps the original id and picture.
      parameters:
      - description: Stage id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Replacement stage
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Replace a stage
      tags:
      - stages
  /stages/{id}/picture:
    patch:
      consumes:
      - multipart/form-data
      description: |-
        Admin only. Uploads the image to object storage and stores its key; the response's `picture` is a presigned URL.
        The image is re-encoded to WebP by a background worker: the response is
        202 Accepted with `pictureStatus` set to PENDING and the previous
        `picture`/`pictureThumb` URLs (or "" on a first upload); poll GET
        /stages/{id} until `pictureStatus` becomes READY or FAILED.
      parameters:
      - description: Stage id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Image file (WebP, AVIF, JPEG, PNG or GIF)
        in: formData
        name: picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.StageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set a stage's picture
      tags:
      - stages
  /stages/{id}/translations:
    get:
      description: |-
        Admin only. Unlike GET /stages/{id}, always returns every
        locale's description at once, for an edit form.
      parameters:
      - description: Stage id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StageTranslationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get every locale's translation for a stage
      tags:
      - stages
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"